fixup! Restructuring, still a work in progress
