+++ lms/djangoapps/lms_course_cache/__init__.py
@@ -1,0 +1,5 @@
+"""
+...
+"""
+# TODO me: Docstrings for all of course_cache/
+import signals  # pylint: disable=unused-import
