--- openedx/core/djangoapps/course_cache/tests/test_blocks.py
@@ -1,9 +1,0 @@
-"""
-...
-"""
-
-course_1 = {
-    'key': 'a1',
-}
-
-# TODO: write this
