+++ openedx/core/lib/course_cache/__init__.py
@@ -1,0 +1,4 @@
+"""
+...
+"""
+# TODO me: Docstrings for all of course_cache/
